backend/requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.1
python-dotenv>=1.0
dj-database-url>=2.2
djangorestframework>=3.15
django-cors-headers>=4.4
